+++ client/beadledom-jaxrs-clientproxy/src/test/scala/com/cerner/beadledom/client/proxy/GenericClientProxySpec.scala
@@ -18,6 +18,7 @@ class GenericClientProxySpec extends FunSpec with MustMatchers with MockitoSugar
         val mockResponse = mock[Response]
         when(mockResponse.readEntity(new GenericType(classOf[String]))).thenReturn("Hello World")
         when(mockResponse.getStatus).thenReturn(200)
+        when(mockResponse.hasEntity).thenReturn(true)
         val underlyingProxy = mock[StandardTestingResource]
         when(underlyingProxy.getGeneric).thenReturn(mockResponse)
 
@@ -51,6 +52,27 @@ class GenericClientProxySpec extends FunSpec with MustMatchers with MockitoSugar
 
         response mustBe a [Response]
         response.asInstanceOf[Response].readEntity(classOf[String]) mustBe "Hello World"
+      }
+
+      it("transform the underlying response with no entity into a GenericResponse") {
+        val mockResponse = mock[Response]
+        when(mockResponse.hasEntity).thenReturn(false)
+        when(mockResponse.getStatus).thenReturn(204)
+        val underlyingProxy = mock[StandardTestingResource]
+        when(underlyingProxy.getGenericWithNoType).thenReturn(mockResponse)
+
+        val proxy = mock[GenericTestingResource]
+
+        val clientProxy = new GenericClientProxy(underlyingProxy)
+
+        val method = classOf[GenericTestingResource].getMethod("getGenericWithNoType")
+        val args: Array[AnyRef] = Array[AnyRef]()
+        val response = clientProxy.invoke(proxy, method, args)
+
+        response mustBe a [GenericResponse[_]]
+        val genericResponse = response.asInstanceOf[GenericResponse[_]]
+        genericResponse.body() mustBe null.asInstanceOf[Void]
+        genericResponse.getStatus mustBe 204
       }
     }
   }